parquet/parquet-py/pyproject.toml
[project]
name = "parquet-py"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Balagopal Kanattil", email = "balukanattil@gmail.com" }
]
dependencies = [
    "thrift>=0.20.0",
    "rich>=13.7.1",
    "python-snappy>=0.7.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
read-example = 'parquet_py:read_example'

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/parquet_py"]
